refactor: move lock to chared api
