+++ cloud/flamingock-cloud-api/src/main/java/io/flamingock/cloud/api/response/LockInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Flamingock (https://www.flamingock.io)
+ * Copyright 2023 Flamingock (https://www.flamingock.io)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,6 @@ public class LockInfo {
     private String acquisitionId;
 
     private long acquiredForMillis;
-
 
     public LockInfo() {
     }
@@ -60,4 +59,5 @@ public class LockInfo {
     public void setAcquiredForMillis(long acquiredForMillis) {
         this.acquiredForMillis = acquiredForMillis;
     }
+
 }
