--- core/flamingock-core-commons/src/main/java/io/flamingock/internal/common/cloud/lock/LockResponse.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2023 Flamingock (https://www.flamingock.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.flamingock.internal.common.cloud.lock;
-
-public class LockResponse {
-
-    private String key;
-
-    private String owner;
-
-    private String guid;
-
-    private long lockAcquiredForMillis;
-
-    public LockResponse() {
-    }
-
-    public String getKey() {
-        return key;
-    }
-
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    public String getOwner() {
-        return owner;
-    }
-
-    public void setOwner(String owner) {
-        this.owner = owner;
-    }
-
-    public String getGuid() {
-        return guid;
-    }
-
-    public void setGuid(String guid) {
-        this.guid = guid;
-    }
-
-    public long getLockAcquiredForMillis() {
-        return lockAcquiredForMillis;
-    }
-
-    public void setLockAcquiredForMillis(long lockAcquiredForMillis) {
-        this.lockAcquiredForMillis = lockAcquiredForMillis;
-    }
-
-}
